feat(frontend): split zIndexes in a new file
