+++ frontend/src/lib/zIndexes.ts
@@ -1,0 +1,12 @@
+export const zIndexes = {
+	appEditorComponents: 1000,
+	secondaryMenu: 1001,
+	splitPanelHandle: 1001,
+	colorInput: 1002,
+	disposables: 1100, // Modals and Drawers
+	svelteSelectOptions: 5000,
+	contextMenu: 6000,
+	draggingComponent: 10000,
+	monacoEditor: 10000,
+	monacoEditorSuggestions: 10001
+}
